accel/amdxdna: split the common DRM core into amdxdna_drm.c
The transport-independent DRM/accel layer -- amdxdna_drm_drv, its file
operations and ioctls, per-client (drm_file) open/close, and the IOMMU SVA
(PASID) binding -- lived in amdxdna_pci_drv.c, so a non-PCI build could not
reach it without pulling in the PCI module entry point.

Move that layer verbatim into a new amdxdna_drm.c, and export amdxdna_sva_fini()
through amdxdna_drv.h so each bus driver's remove path can unbind clients.
amdxdna_pci_drv.c keeps only the PCI specifics (id tables, probe/remove, PM, FLR
and SR-IOV).

A follow-up Kbuild change compiles amdxdna_drm.o into both builds; the default
PCI build is otherwise unchanged.
