+++ drivers/accel/amdxdna/amdxdna_drv.h
@@ -277,6 +277,9 @@ void amdxdna_iommu_free(struct amdxdna_dev *xdna, size_t size,
 int amdxdna_dma_map_bo(struct amdxdna_dev *xdna, struct amdxdna_gem_obj *abo);
 void amdxdna_dma_unmap_bo(struct amdxdna_dev *xdna, struct amdxdna_gem_obj *abo);
 
+/* Unbind a client's IOMMU SVA/PASID; defined in the shared amdxdna_drm.c. */
+void amdxdna_sva_fini(struct amdxdna_client *client);
+
 static inline bool amdxdna_iova_on(struct amdxdna_dev *xdna)
 {
 	return !!xdna->domain;
